Rename Gradle plugin package
