+++ arrow-libs/plugins/build.gradle.kts
@@ -10,18 +10,18 @@ repositories {
 dependencies {
   implementation(gradleApi())
   implementation(kotlin("stdlib"))
-  implementation("org.jetbrains.dokka:dokka-gradle-plugin:1.4.30")
+  implementation("org.jetbrains.dokka:dokka-gradle-plugin:1.5.0")
 }
 
 gradlePlugin {
   plugins {
     register("documentation") {
       id = "documentation"
-      implementationClass = "io.github.nomisrev.DocumentationPlugin"
+      implementationClass = "arrow.gradle.DocumentationPlugin"
     }
     register("mpp-publish") {
       id = "mpp-publish"
-      implementationClass = "io.github.nomisrev.MppPublishingPlugin"
+      implementationClass = "arrow.gradle.MppPublishingPlugin"
     }
   }
 }
