+++ arrow-libs/plugins/src/main/kotlin/arrow/gradle/MppPublishingPlugin.kt
@@ -1,4 +1,4 @@
-package io.github.nomisrev
+package arrow.gradle
 
 import org.gradle.api.Plugin
 import org.gradle.api.Project
